+++ docs/ProcessManager.Orchestrations.Abstractions/ReleaseNotes/ReleaseNotes.md
@@ -1,5 +1,9 @@
 # ProcessManager.Orchestrations.Abstractions Release Notes
 
+## Version 2.4.3
+
+- Update NuGet package dependencies.
+
 ## Version 2.4.2
 
 - Add `RequestMeasurementsAcceptedV1`